verify the sign with msgMeta
