+++ core/src/main/java/tech/pegasys/web3signer/core/FilecoinRunner.java
@@ -12,6 +12,8 @@
  */
 package tech.pegasys.web3signer.core;
 
+import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
+import static com.fasterxml.jackson.databind.MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS;
 import static com.fasterxml.jackson.databind.MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES;
 import static io.vertx.core.http.HttpHeaders.CONTENT_TYPE;
 import static tech.pegasys.web3signer.core.service.http.handlers.ContentTypes.JSON_UTF_8;
@@ -39,8 +41,6 @@ import tech.pegasys.web3signer.core.signing.filecoin.FilecoinNetwork;
 import java.util.Collection;
 import java.util.List;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.arteam.simplejsonrpc.server.JsonRpcServer;
 import io.vertx.core.Vertx;
@@ -85,9 +85,9 @@ public class FilecoinRunner extends Runner {
     final FcJsonRpc fileCoinJsonRpc = new FcJsonRpc(fcSigners, fcJsonRpcMetrics);
     final ObjectMapper mapper =
         new ObjectMapper()
-            .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS)
+            .enable(ACCEPT_CASE_INSENSITIVE_ENUMS)
             .enable(ACCEPT_CASE_INSENSITIVE_PROPERTIES)
-            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+            .disable(FAIL_ON_UNKNOWN_PROPERTIES)
             .registerModule(new FilecoinJsonRpcModule());
     final JsonRpcServer jsonRpcServer = JsonRpcServer.withMapper(mapper);
